+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/usecase/CancelMessage.kt
@@ -11,7 +11,6 @@ public interface CancelMessage {
      * Cancels the message of "ephemeral" type. Removes the message from local storage.
      * API call to remove the message is retried according to the retry policy specified on the chatDomain
      * @param message the message to send
-     * @return A call object with Message as the return type
      * @see io.getstream.chat.android.livedata.utils.RetryPolicy
      */
     public operator fun invoke(message: Message): Call<Boolean>