+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/list/adapter/viewholder/ChannelViewHolder.kt
@@ -146,7 +146,6 @@ public class ChannelViewHolder @JvmOverloads constructor(
         channelClickListener: ChannelListView.ChannelClickListener
     ) {
         avatarView.apply {
-            setStyle(style.avatarStyle)
             setChannelData(channel)
             setOnClickListener {
                 when {
@@ -243,9 +242,9 @@ public class ChannelViewHolder @JvmOverloads constructor(
         }
 
         binding.apply {
-            channelNameLabel.setTextSizePx(style.channelTitleText.size.toFloat())
-            lastMessageLabel.setTextSizePx(style.lastMessage.size.toFloat())
-            lastMessageTimeLabel.setTextSizePx(style.lastMessageDateText.size.toFloat())
+            channelNameLabel.setTextSizePx(style.channelTitleTextSize)
+            lastMessageLabel.setTextSizePx(style.lastMessageSize)
+            lastMessageTimeLabel.setTextSizePx(style.lastMessageDateTextSize)
         }
     }
 }